remove 'clear' option
We won't need it anymore because we're changing how reports are generated and stored in this version , it's unnecessary and redundant. The user shouldn't have to run a terminal command to clear the report. If a generated report is not useful, they can simply delete the file. Also, we will generate 1 report file for simplicity — because let's face it, nobody has the time to go into reach folder and read an individual report file per code file. Moreover, we don't need reports per code file - that's most likely intermediate processing that the user should not see. More on this in the release notes.